refactor(spark): move backup success screen to shared spark-onboarding level
