+++ app/screens/spark-onboarding/index.ts
@@ -2,4 +2,4 @@ export { SparkBackupMethodScreen } from "./backup-method-screen"
 export { SparkBackupAlertsScreen } from "./manual-backup/backup-alerts-screen"
 export { SparkBackupPhraseScreen } from "./manual-backup/backup-phrase-screen"
 export { SparkBackupConfirmScreen } from "./manual-backup/backup-confirm-screen"
-export { SparkBackupSuccessScreen } from "./manual-backup/backup-success-screen"
+export { SparkBackupSuccessScreen } from "./backup-success-screen"
